@@ -1,7 +1,9 @@
-from typing import (
-    TYPE_CHECKING,
+from collections.abc import (
     Iterable,
     Sequence,
+)
+from typing import (
+    TYPE_CHECKING,
     TypeVar,
 )
 
