@@ -1,7 +1,8 @@
+from collections.abc import (
+    Sequence,
+)
 from typing import (
     Any,
-    Sequence,
-    Tuple,
     Union,
 )
 
@@ -72,7 +73,7 @@ class Bitlist(BitfieldCompositeSedes[BytesOrByteArray, bytes]):
     # Deserialization
     #
     @to_tuple
-    def deserialize(self, data: bytes) -> Tuple[bool, ...]:
+    def deserialize(self, data: bytes) -> tuple[bool, ...]:
         # Length of data should be larger than 1
         if len(data) < 1:
             raise DeserializationError(
@@ -110,7 +111,7 @@ class Bitlist(BitfieldCompositeSedes[BytesOrByteArray, bytes]):
 
     def get_hash_tree_root_and_leaves(
         self, value: Sequence[bool], cache: CacheObj
-    ) -> Tuple[Hash32, CacheObj]:
+    ) -> tuple[Hash32, CacheObj]:
         root, cache = merkleize_with_cache(
             pack_bits(value), cache=cache, limit=self.chunk_count
         )
